Minor fixes in xmonad build script

+++ xmonad/.xmonad/build.sh
@@ -3,13 +3,11 @@
 cd $HOME/.xmonad
 
 # TODO: Check and compare version
-if [[ $* == *--update* ]]; then
-  if [ -d .cabal-sandbox ]; then
-    if [ -d .cabal-sandbox-prev ]; then
-      rm -rf .cabal-sandbox-prev
-    fi
-    mv .cabal-sandbox .cabal-sandbox-prev
+if [[ $* == *--update* && -d .cabal-sandbox ]]; then
+  if [ -d .cabal-sandbox-prev ]; then
+    rm -rf .cabal-sandbox-prev
   fi
+  mv .cabal-sandbox .cabal-sandbox-prev
 fi
 
 # Check if cabal sandbox exists, otherwise create it and install xmonad
@@ -23,7 +21,7 @@ fi
 
 # Build the xmonad config
 cabal build
-if [ -f xmonad-x86_64-linux ]; then
+if [[ -f dist/build/myxmonad/myxmonad && -f xmonad-x86_64-linux ]]; then
   mv xmonad-x86_64-linux xmonad-x86_64-linux.prev
 fi
 mv dist/build/myxmonad/myxmonad xmonad-x86_64-linux
